requirements.txt
asgiref ==3.3.4
certifi==2022.12.7
Django==3.2.15
django-simpleui==2023.12.12
djangorestframework==3.14.0
mysqlclient==2.1.0
psycopg2==2.8.6
PyMySQL==1.1.0
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.7.1
wincertstore==0.2
